(Almost) Finished statues!
The stateus are almost done! All that needs to be done is to allow
players to choose which player to make a statue of, to make sure to get
the shoulders right, and to get rid of my horrific skin! Four bug fixes
are that the right leg is the right size, that the right arm is the
right size, that the head is the right size, and that the top of the
head appears on top of the head (instead of the face appearing there).